Refactor rule/offender representations
Things are getting complicated.

So the problem is that each rule may have a description and solution
that depend on variables from an offender, but there are only ever a
single instance of each rule, so they can’t hold state.

This solution is more complicated than I really like, but works for now
and allows you to provide everything related to representation/output
in each rule module.