Modify compliance algorithm
It’s still not great, but I think this yields more indicative scores.

If you have a lot of files, each with only a few violations, you’ll get
a bad score. If you have a single file with lots of violations, you’ll
still get a bad score, but better than the previous case.