Add a program-wide way of printing to the appropriate outputs
By providing the concepts of a diagnostic and a result, we don't have
to keep reminding ourselves which buffer a text should go to. Just
print using one of the two print* methods.

Also adds the StandardReporter as a separate thing from the base
Reporter, and a --verbose program argument for enabling diagnostic
messages.