Refactor suppression of similar violations
This also has the side-effect of removing suppression of identical
suggestion/solutions. I think it’s worth it, though, as the code was
increasingly difficult to comprehend.

This puts more of the general workload into the base Reporter, making
each specialized reporter more focused on just forming the output.

See #8