Make path per result appear first thing, but truncated
The length of this path i'm not so sure about... It needs to be long
enough to figure out which file it points to, but also short enough
that there's room for the violation description.

It's important that the path component is a constant width (even if
shorter- in which case, it is padded), so that entire report is aligned.