Require original text
These rules require //-line comments to remain in text, as such, they
must be handed the original text to work.

This, however, also has the downside that false-positives may occur on
#include statements that are blocked out by a comment, or are found
inside a block-style comment.

This could be improved if rules could specify the degree of
“originality” they require of the text; e.g. they could specify
`with_block_comments=True`, or `with_line_comments=True`, so in this
case, the rules could be handed a text with block style comments
stripped, but line comments remaining- eliminating possibility of at
least some false-positives.