Relax severity of func-too-long and too-many-funcs
These rules should be used more like a guideline than law.

A function may be perfectly fine >40 lines, and a file may be perfectly
fine >7 functions; it all comes down to making a conscious decision
about it. Less code is (always) better- but if the code is just split
into obscurity to satisfy an arbitrary rule, then it is not.