Replace const-on-right with a more general rule
This makes `const-on-right` catch more cases anywhere in a source file;
not limited to function parameters.

It's a better and (much) simpler solution that fixes issues found in
#40 #41

A few downsides:

 * It has worse fix-it context output (can be mitigated)
 * Incorrectly reports the following as a violation:

```
int
const a = 1;
```

This is because the pattern essentially just looks for `const`s with
blank on its left.