Refactor stripping of string literals
I’m not sure why I thought it was important to “retain line
terminators” as these would be *inside* the literal anyway, and thus
have no significance for actual line breaking.

Additionally, this commit changes how the strings are stripped. Instead
of just removing the entire string, this solution leaves literal
markers behind (e.g. “ or ‘). This leaves the code in a working state,
which should help avoid matching issues.