Refactor violation collection
This commit removes text, filename and extension parameters from the
`collect` function and replaces them with a single `CheckFile`
parameter.

This object represents a source file being checked, and allows for more
options on what to match and grab snippets from. Previously, it would
be confusing what `text` actually represented; was it stripped, or was
it the original text body? Now it is explicitly stated in the rule
implementation.

See #35