Match rule templates with rule classes
The build script now automatically looks up all rule classes and tries
to match them up with the provided templates- if anything is off, e.g.
naming, the script can now let us now.

This opens up possibilities for grabbing stuff directly from the code,
without having to type it all up. For example, it might be useful to
have the rule descriptions in the docs along with the current texts.