Add test helper functions for expected violations
These will allow much easier test setups, essentially all a test needs
to do is supply some chunks of text and indicate in that text where
violations are expected to occur.

I saw this being used on the SwiftLint project and thought it might
ease the test burden here as well.