Remove too-many-funcs
This rule is too arbitrary and i've found it to rarely be a useful indicator of code smell.

A source file could easily violate this rule while still being logically sound and clean, and in such a case, getting this violation would likely just be counter-productive.

I do still think keeping number of functions low is a thing to strive for, but the "max number of functions" is too situational to be put into a static rule.

I think the better metric is number of lines in file (see file-too-long), and coupled with func-too-long, you will land at a maximum number of functions anyway. Though these are absolutely more objectively defensible metrics, they both ultimately suffer from the same issue; sometimes more *is* better.